feat(websocket): added the websocket server-client implementation
