+++ data-plane/Cargo.toml
@@ -70,10 +70,13 @@ drain = { version = "0.2", features = ["retain"] }
 duration-string = { version = "0.5.3", features = ["serde"] }
 futures = "0.3.31"
 futures-timer = "3.0.3"
+fastwebsockets = { version = "0.10.0", features = ["upgrade", "unstable-split"] }
 h2 = "0.4.7"
 headers = "0.4.1"
 hex = "0.4.3"
 http = "1.2.0"
+http-body-util = "0.1.3"
+hyper = { version = "1.8.1", features = ["http1", "server", "client"] }
 hyper-rustls = { version = "0.27", features = ["http2", "aws-lc-rs"], default-features = false }
 hyper-util = "0.1.10"
 indexmap = "2"
@@ -123,6 +126,7 @@ test-fork = "0.1.3"
 thiserror = "2.0.9"
 tokio = "1.52.2"
 tokio-retry = "0.3"
+tokio-rustls = "0.26.4"
 tokio-stream = "0.1"
 tokio-test = "0.4.4"
 tokio-util = "0.7.14"
