+++ data-plane/core/config/Cargo.toml
@@ -10,18 +10,22 @@ name = "slim_config"
 
 [[bin]]
 name = "generate-schema"
-path = "src/grpc/schema/generate_schema.rs"
+path = "src/schema/generate_schema.rs"
 
 [dependencies]
 agntcy-slim-auth = { workspace = true }
 agntcy-slim-version = { workspace = true }
 async-trait = { workspace = true }
 base64 = { workspace = true }
+bytes = { workspace = true }
 display-error-chain = { workspace = true }
 drain = { workspace = true }
 duration-string = { workspace = true }
+fastwebsockets = { workspace = true }
 futures = { workspace = true }
 http = { workspace = true }
+http-body-util = { workspace = true }
+hyper = { workspace = true }
 hyper-rustls = { workspace = true }
 hyper-util = { workspace = true, features = ["client-proxy-system"] }
 lazy_static = { workspace = true }
@@ -38,6 +42,7 @@ serde_yaml = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-retry = { workspace = true }
+tokio-rustls = { workspace = true }
 tokio-stream = { workspace = true }
 tokio-util = { workspace = true }
 tonic = { workspace = true }
@@ -60,4 +65,3 @@ agntcy-slim-testing = { workspace = true }
 rand = { workspace = true }
 serde_json = { workspace = true }
 tracing-test = { workspace = true }
-
