+++ data-plane/core/config/src/schema/client-config.schema.json
