+++ data-plane/core/config/src/schema/generate_schema.rs
@@ -1,6 +1,6 @@
 use schemars::{JsonSchema, schema_for};
-use slim_config::grpc::client::ClientConfig;
-use slim_config::grpc::server::ServerConfig;
+use slim_config::client::ClientConfig;
+use slim_config::server::ServerConfig;
 use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
@@ -10,7 +10,7 @@ fn write_schema<T: JsonSchema>(file_name: &str) {
     let schema_json = serde_json::to_string_pretty(&schema).unwrap();
 
     let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    path.push(format!("src/grpc/schema/{}", file_name));
+    path.push(format!("src/schema/{}", file_name));
 
     let mut file = File::create(&path).unwrap();
     file.write_all(schema_json.as_bytes()).unwrap();
