+++ data-plane/core/config/src/schema/server-config.schema.json
