+++ data-plane/core/config/src/tls/client.rs
@@ -286,6 +286,9 @@ impl TlsClientConfig {
             return Ok(None);
         }
 
+        // Ensure rustls has a process-level crypto provider before any config builders run.
+        crate::tls::provider::initialize_crypto_provider();
+
         // Resolve TLS version
         let tls_version = match self.config.tls_version.as_str() {
             "tls1.2" => &TLS12,
