+++ data-plane/core/config/src/websocket.rs
@@ -1,0 +1,6 @@
+// Copyright AGNTCY Contributors (https://github.com/agntcy)
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod client;
+pub mod common;
+pub mod server;
