+++ data-plane/core/config/tests/e2e.rs
@@ -4,7 +4,7 @@
 use tonic::{Request, Response, Status, metadata::KeyAndValueRef};
 use tracing::info;
 
-use slim_config::grpc::client::ClientConfig;
+use slim_config::client::ClientConfig;
 use slim_config::testutils::helloworld::greeter_server::Greeter;
 use slim_config::testutils::helloworld::{HelloReply, HelloRequest};
 
@@ -93,10 +93,10 @@ mod tests {
     // use slim_config_grpc::headers_middleware::SetRequestHeader;
     use slim_auth::jwt::{Key, KeyData};
     use slim_auth::traits::Signer;
-    use slim_config::grpc::{client::ClientConfig, server::ServerConfig};
     use slim_config::testutils::helloworld::HelloRequest;
     use slim_config::testutils::helloworld::greeter_client::GreeterClient;
     use slim_config::testutils::helloworld::greeter_server::GreeterServer;
+    use slim_config::{client::ClientConfig, server::ServerConfig};
     use slim_testing::utils::setup_test_jwt_resolver;
     #[cfg(unix)]
     use {
@@ -146,7 +146,7 @@ mod tests {
         });
 
         // create a client using the channel
-        let channel = match client_config.to_channel().await {
+        let channel = match client_config.to_grpc_channel().await {
             Ok(ch) => ch,
             Err(e) => return Err(Box::new(e)),
         };
@@ -199,7 +199,7 @@ mod tests {
         yield_now().await;
 
         // Use the config-driven client to connect over the Unix socket
-        let channel = client_config.to_channel().await?;
+        let channel = client_config.to_grpc_channel().await?;
         let mut client = GreeterClient::new(channel);
 
         let request = tonic::Request::new(HelloRequest {
@@ -314,7 +314,7 @@ mod tests {
         // create a new client with wrong credentials
         let channel = client_config
             .with_auth(auth_wrong_client_config)
-            .to_channel()
+            .to_grpc_channel()
             .await?;
 
         let mut client = GreeterClient::new(channel);
