+++ data-plane/core/datapath/Cargo.toml
@@ -18,7 +18,9 @@ agntcy-slim-tracing = { workspace = true }
 agntcy-slim-version = { workspace = true }
 display-error-chain = { workspace = true }
 drain = { workspace = true }
+fastwebsockets = { workspace = true }
 h2 = { workspace = true }
+http = { workspace = true }
 opentelemetry = { workspace = true, optional = true }
 parking_lot = { workspace = true }
 prost = { workspace = true }
