+++ data-plane/core/datapath/src/websocket.rs
@@ -1,0 +1,6 @@
+// Copyright AGNTCY Contributors (https://github.com/agntcy)
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod stream;
+
+pub(crate) use stream::spawn_transport_tasks;
